Update Registry Manager to support multi-tenancy and OpenSearch Serverless (#73)
* matches new registry-common API

The only think to note here, is that arg value changes but everything else remains the same. Instead of https://localhost:9200 will be app://localhost.xml.

* new commands

Allow the user to know which connections are built into the artifacts and then fetch them (print them to standard out). Once printed out they can be saved to a file and modified. The file can them be used using file:// rather than app://

* cleaned up get and search

Should have separated all of the get/search parts and everything that is es or os is now in registry-common.