handle leap secondss
When leap seconds are of concern, shift them to a Java valid time then convert to standard format then add the 60 second leap back into it. This does a couple of things. First, it avoids the java annoyance of not handling the 60 second leap. Two, it uses the PDS4 regexp for ASCII date times to know when it is appropriate. Third, it keeps the 60 second leap appropriately in the string format.