correctly handle nested structures
The mapping handling was built on the concept if the database being flat. Now that it is not, had to back up and redo it. These changes should work with flat as well as with nested structures. Basically, remove the heritage from any leaf node but its direct parent and then look for it in the registry-dd. It seems to work as all of the test labels in registry-ref-data are being found and ingested with the correct types. Also, missing items are using the structured approach too. Again, changes should be backward compatible as flat is a unique case of nested.