harvest_244: change the registry frame work (#144)
* updated manger for new structured db

Started with structured and versions. Structured is added to the base registry name from the XML file. It should prevent older dbs from mucking with the newer if all goes well. At least for harvest as harvest has to take the name and convert it to an ENUM. The manager tools are more flexible, but it should cause most some pain in getting around the structured suffix.

Then added versioning. The tool now builds a version table and checks its own version against what the DB wants as a minimum. Verified its operation. The create-registry will skip versioning checks because the table has not been created yet.

* silly sonar fix

* better verbage

* add check for tool not present

* more refactoring as I learn more

* cleaned up versioning some more

* add a command to update versions in the registry

* update version

* clean up for new DOMing

* change / to .

* clean up alerts

* update for harvest 261

* update date to string (#153)

Part 1 was to fix the configuration. Had to make same change in harvest. Really should have only one data-dic-types.cfg in registry-common and not two want to be the same in regsistry-mgr and harvest.

Part 2: ditch the data-dic-data.jar because it is old and defines dates as opensearch date. Need to rebuild it with new code on this branch.

Part 3: fix LDD loader to set index correctly.

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>

---------

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>